backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
werkzeug==2.3.7
gunicorn==21.2.0
redis==5.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1
bcrypt==4.0.1
PyJWT==2.8.0
bleach==6.1.0